Renaming batchutilities package to batchmanager
