+++ core/sdk-core/src/main/java/software/amazon/awssdk/core/internal/batchmanager/ScheduledFlush.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.core.internal.batchutilities;
+package software.amazon.awssdk.core.internal.batchmanager;
 
 import java.util.concurrent.ScheduledFuture;
 import software.amazon.awssdk.annotations.SdkInternalApi;
