+++ core/sdk-core/src/test/java/software/amazon/awssdk/core/internal/batchmanager/ScheduledFlushTest.java
@@ -13,7 +13,7 @@
  * permissions and limitations under the License.
  */
 
-package software.amazon.awssdk.core.internal.batchutilities;
+package software.amazon.awssdk.core.internal.batchmanager;
 
 import java.util.ArrayList;
 import java.util.Iterator;
