+++ services/sqs/src/it/java/software/amazon/awssdk/services/sqs/BatchManagerSqsIntegrationTest.java
@@ -35,12 +35,12 @@ import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
-import software.amazon.awssdk.core.internal.batchutilities.BatchAndSend;
-import software.amazon.awssdk.core.internal.batchutilities.BatchKeyMapper;
-import software.amazon.awssdk.core.internal.batchutilities.BatchManager;
+import software.amazon.awssdk.core.internal.batchmanager.BatchAndSend;
+import software.amazon.awssdk.core.internal.batchmanager.BatchKeyMapper;
+import software.amazon.awssdk.core.internal.batchmanager.BatchManager;
 import software.amazon.awssdk.core.BatchOverrideConfiguration;
-import software.amazon.awssdk.core.internal.batchutilities.BatchResponseMapper;
-import software.amazon.awssdk.core.internal.batchutilities.IdentifiableMessage;
+import software.amazon.awssdk.core.internal.batchmanager.BatchResponseMapper;
+import software.amazon.awssdk.core.internal.batchmanager.IdentifiableMessage;
 import software.amazon.awssdk.services.sqs.model.CreateQueueRequest;
 import software.amazon.awssdk.services.sqs.model.SendMessageBatchRequest;
 import software.amazon.awssdk.services.sqs.model.SendMessageBatchRequestEntry;
